requirements.txt
pandas
matplotlib
seaborn
scikit-learn
psycopg2-binary
joblib
python-dotenv   